pyproject.toml
[tool.poetry]
name = "explorer2go"
version = "0.1.0"
description = ""
authors = ["Danilo da Conceição Nascimento <daconnas.dcn@gmail.com>"]
readme = "README.md"
packages = [{include = "explorer2go"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-extensions = "^3.2.3"
gunicorn = "^21.2.0"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
pillow = "^10.0.0"
django-storages = "^1.13.2"
boto3 = "^1.28.15"
django-mathfilters = "^1.0.0"
fontawesomefree = "^6.4.0"
whitenoise = "^6.5.0"
django-import-export = "^3.2.0"
django-colorfield = "^0.9.0"
sentry-sdk = "^1.29.2"
django-tenants = "^3.5.0"
psycopg2-binary = "^2.9.7"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
taskipy = "^1.12.0"
blue = "^0.9.1"
ruff = "^0.0.280"
httpx = "^0.24.1"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
parsel = "^1.8.1"
locust = "^2.16.1"


[tool.poetry.group.infra.dependencies]
terraform-local = "^0.12"
awscli-local = "^0.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79

[tool.pytest.ini_options]
pythonpath = "."
python_files = "tests.py test_*.py *_tests.py"

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = 'gunicorn'
dev = 'gunicorn --reload'
migrate = 'python manage.py migrate_schemas'
test = 'pytest --disable-warnings -s -x --cov=explorer2go -vv --ds=tests.settings'
post_test = 'coverage html'

[tool.coverage.run]
omit = [
    "**/tests/**",
    "**/migrations/**",
]